30/10/2017
changed the 1 billion count simulation. It had x2 and y2 terms in the A matrix of CMacDonald2017.m file. This translates to the velocity of mass 1 and mass 2. So the only reason I can see them being in the A matrix was that I was implementing some kind of damping term.