5 frequencies for different no. of cells
Created code for 2,5,10U systems exposed to the 5 different frequencies. Might need to adjust the frequencies per system as some resonances that appear in the 2 U system do not appear in the same place as the 5 and 10U system. Also, might need to adjust the sensor location. Current sensor location is the second unit primary mass. Should be the last unit's primary mass. Should investigate.